Option to use system locale for character conversion.
Solves the problem described in the following report:
Problem Converting Letters in Filename from Uppercase to id3 Fields in
Lowercase with Dotted and Dotless I
To replicate the problem use a file that has I and İ in the filename. Set
convert to lowercase in settings and try to get the id3 field from the
filename. İ is converted to I and I is converted to i. Normally İ should be
converted to i and I to ı.

Using the locale has to be switched on explicitly as it may not be desired
if the tags or filenames are not in the locale language. The language package
must be installed, QT >= 4.8.0 is required. Kid3 can be started with a
specific locale, using for example "LANG=tr_TR.UTF-8 ./kid3-qt".