Supply translation file for English.
Even if this file does not contain any translations at all, it is not
useless: If multiple languages are available (e.g. separated by colons
in the LANGUAGE variable), English would be skipped if no language
file is available. For example, if LANGUAGE=en:de and no language file
for en is available, de will be used even if en is preferred.