Qt-only: Try to load translations for UI language files respecting English.
QTranslator::load(const QLocale&, ...) has the problem that English is
skipped because there is no English translation file (providing an
English translation file for Kid3 does not help as there is no qt_en.qm
for Qt, which will give non-English strings in Qt dialogs). The problem
occurs if the first language is not English (that special case was handled)
but a language without a translation file, e.g. Swedish, and English is
the fallback language. In this case, English will be skipped and the
next fallback language is loaded, e.g. using "LANGUAGE=sw:en:fr kid3-qt"
sw is not found, en is skipped and fr is loaded giving a French translation
instead of English. Now this is fixed by using
QTranslator::load(const QString&, ...) explicitly for every UI language.