Cache more information for closed TagLib files.
Only a limited number of TagLib files are open at the same time to reduce
the number of open file descriptors. Some information could not be read
for closed files, e.g. isTagV1Supported(), getDetailInfo(), getDuration(),
getFileExtension(). Now the needed information is cached, so that it is
available. The problem occurred when getting the duration of many files in
a batch import.