Use QLatin1String, QLatin1Char to avoid implict conversions from char* to QString.
QString(const char*) is equivalent to QString::fromLatin1() in Qt4 (if no
codec is set), however in Qt5, it is QString::fromUtf8() which can lead to
unnoticed bugs. Strings from ID3v1 tags in Mp3File were not converted
correctly with Qt5, so QT_ASCII_CAST_WARNINGS is now defined and all
warnings are fixed.