Fix lintian errors by making the private libraries versioned.
lintian reported pkg-has-shlibs-control-file-but-no-actual-shared-libs
caused by "dpkg-shlibdeps: warning: Can't extract name and version from
library name" for libkid3-core.so and libkid3-core.so. Now these two
libraries have a version so that it can be extracted from the library
name. However, they are still private and will not keep binary compatibility.
Additionally, lintian reported postinst-has-useless-call-to-ldconfig and
postrm-has-useless-call-to-ldconfig because the private libraries are in
a private directory. Now "dh_makeshlibs --noscripts" is used to avoid
generation of postinst and postrm scripts.

Unfortunately it does not seem to be possible to create unversioned
shared libraries without complaints from Debian's lintian. Removing the
SONAME with the NO_SONAME cmake property will result in a full path to
the library and the dynamic loader will not find the library.