Remove file filter if necessary to open a dropped file.
The file filter of the file list is set from the file type filter selected
in the open dialog. Only files which match this wildcard pattern are
displayed in the file list. If a file with an extension which is not in
this filter is dropped into Kid3, it could not be opened. Now the file
filter is removed in such a case, so that dropped files can always be
opened. This problem occurred when a user wanted to drop an Opus file, but
this file type was not supported before and thus not in the user's
"all supported files" filter.