Revival of HTML parsing for Discogs import.
Unfortunately, using the Discogs API is not practical anymore.
They now require OAuth authentication for the whole API after
having introduced it first for images. Probably they would
find another means to block it once OAuth authentication would
be implemented, so I return to HTML parsing and the Sisyphean task
of adapting it all the time.