Fix crash when TagLib file is saved with changes in tags and file name.
FileIOStream::tell() crashes after changed tags have been save, the file
is renamed and reread because the file stream cannot be opened because
it still uses the old file name as returned by the file system model which
is not updated yet. Now it is checked if the file is really open.