Set root of file list model before selection.
If the selection is set first, the root is not set correctly when opening
a network share on Mac OS X. To reproduce it, open a file on a network share,
set configuration to open the last file on startup and restart Kid3.
Then the root directory / is shown in the file list, not the directory of the
selected file.