Revert file name modification if tagged file is not writable.
Changing the permissions does not work if both the tags and
the file name are changed on a read-only file. In such a case
the changed file name is reverted so that changing the
permissions and saving the changed tags will work. The file
name change is lost, but this is less severe than losing all
tag changes.