Avoid blocking when start filtering after filtering out a lot of files.
When a lot of files are filtered out,
QSortFilterProxyModel::invalidateFilter() is extremely slow (probably
depending on the source model). In this case, I measured 3s for 3000
files, 8s for 5000 files, 54s for 10000 files, and too long to wait for
more files. If such a case is detected, the file system model is
recreated in order to avoid calling invalidateFilter().

The blocking could be reproduced by filtering out a lot of files, then
aborting the filter and restarting it.