Fix wrong depth 32 instead of 24 stored in METADATA_BLOCK_PICTURE.
When adding a JPEG picture to an Opus file and then running opusinfo
on it, the following warning is reported:

WARNING: Mismatched picture parameters in METADATA_BLOCK_PICTURE comment
3 (stream 1): depth declared as 32 but appears to be 24

The problem is that QImage::depth() reports 32 whereas
QImage::bitPlaneCount() reports 24, according to the documentation
"This is different from (i.e. smaller than) the depth when the image
format contains unused bits." Now bitPlaneCount() is used instead of
depth() to avoid such warnings.