Fix deletion of multiple METADATA_BLOCK_PICTURE picture frames.
When an Opus file contains multiple pictures and they are deleted using
the Delete button starting with the first, the last could not be
removed. The file had to be saved before being able to delete the last
picture.
This is now fixed by correctly updating the picture indexes when a
picture frame is deleted.