Support atoms purl, egid, cmID, xid with Mp4v2Metadata plugin.
With TagLib, only xid is supported yet. The composer ID (cmID) would
need explicit parsing as UInt (like "atID") and podcast URL (purl) and
podcast GUID (egid) would be text but have the flag 0 (when generated
with AtomicParsley) which is ignored by TagLib's parseText() because
it expects a flag of 1.