Reenable drop to whole window, not only right half.
This reverts part of commit "Drag files from file list to other
applications, image files to embed". It also disables internal drops of
files so that draging a file from the file list to itself does not
open its directory. But it is still possible to drop image files from
the file list and to drop items to a playlist window.