FileSystemModel: Remove watched path when directory is removed or renamed.
In QFileSystemModel, watched paths for directories are only added but never
removed. This leaks resources and leads to wrong behavior when a directory
is renamed. There are watches for the old and the new name, but when
a file is added or removed in the new directory, only the watch for
the old path is activated, the change is not detected and the directory
is not updated.