Avoid jumping around frames while deleting.
When a frame is deleted using the "Delete" button, it can occur that the
current row changes, which is annoying when deleting multiple frames.
This situation can be reproduced with two files where the first file
does not have all the frames which the second file has. If now both files
are selected and one of the frames which only exists in the second file
is deleted, the cursor jumps to a frame which exists in the first file.

This commit fixes this behavior. Additionally, the cursor is also restored
when adding or editing frames in multiple files.