Fix a few details in last commit
For a few things, I think that they were not correct. Other things
are a matter of taste, I prefer if sentences mentioning GUI elements
are not interrupted by punctuation marks therefore such characters
are omitted even if they are part of the button or menu item.
Please avoid using entities which have to be translated.