Add keyboard shortcuts to the sections for editing without the mouse
The standard keyboard shortcuts for Back and Forward can be used to navigate
to other sections, Delete to delete the current frame, Copy and Paste to
copy and paste the section data. Shift-Delete removes the section,
Shift-Ctrl-V pastes data from another seciton, and F2 can be used to edit
the current frame.