Support display names and non user defined frames for custom frames
The custom frames are shown with the translated display names, and
such names can also be used when defining custom frames and will
be translated to internal names unless starting with '!'.
It is also possible to use non user defined (TXXX for ID3) or
free form (MP4) frames for custom frames.