Fix folder renaming when filename format is used
When the option "Filename Format/Use for playlist and folder names" is
activated and the corresponding "String replacement" replaces the '/'
character, it is no longer possible, to rename a folder to multiple
path components (e.g. using the "%{artist}/%{"["year"] "}%{album}"
format in "Tools/Rename Folder").

Now the format is applied separately for each path component and
the '/' characters are preserved.

BUG: 447795