Fix crash when stale tags are read after being stripped
This crash can happen when standard tags are displayed in the
file list (enabled via context menu) and their values are updated
just after stripping their tag. It is enough to strip the tag
using the "Remove" button and then saving. Another way to
reproduce it is using the "Rewrite Tags" action from the file
list context menu, as described in the bug report.

Saving of the stripped tag in TagLibFile::writeTags() called
markTagUnchanged() before setting the corresponding tag to null.
This call then invoked FileProxyModel::onFileModificationChanged(),
emit dataChanged(), further TaggedFileSystemModel::data() to update
the file list and then TagLibFile::getFrame() which tried to read
from the stale tag causing the crash.

Now the state of the tag is updated before the change notification
is emitted so that the subsequent reading of the tags works correctly.

BUG: 447798