Fix another crash when stale tags are read after being stripped
When tags are displayed in the file list, then Tag 1 is removed
from an MP3 file, and then "File/Revert" is invoked. Kid3 crashes
while accessing a stale tag in TagLibFile::getFrame(). This
is invoked by markTagUnchanged() called in TagLibFile::readTags()
while not all tags have been set to null.