TagLib: Allow editing MP4 atoms with four letter names
If an MP4 file contains free form atoms with four letter names
(e.g. "----:com.apple.iTunes:cdec"), it was assumed that this
is a "normal" atom and not a free form atom and it was not
correctly looked up and could neither be edited nor deleted.
Now it is checked if such a free form atom exists.
By the way: Such four letter atom can be added by prefixing them
with a colon, e.g. using ":cdec" for the example above.