Id3libMetadata: Avoid crash when trying to create unsupported frame
When the Id3libMetadata plugin is activated and the Compilation frame is
available via the Quick Access Frames, selecting multiple MP3 files, then
selecting the Compilation frame and clicking Edit will lead to a crash
because such a frame is not supported by id3lib, cannot be created
and a null pointer is dereferenced. Now the frame pointer is checked
before accessing it.

BUG 466194