KDE: Avoid excessive width of file dialog
The file dialog which appears for File/Open on KDE has excessive width,
which seems to depend on the long "All Supported Files" filter.
One solution would be not to use QFileDialg::DontUseNativeDialg,
however, the native KDE dialog which is then used does not seem to work
correctly with flatpak, thus the QFileDialog::HideNameFilterDetails
option is used to avoid the excessive width.