QML: Fix isStandalone() when running non-interactively with kid3-cli
In QML scripts, isStandalone() is used to check if the script is not
returning to the Kid3 UI and thus it has to save modified files itself
as there is no user which can do that interactively afterwards.
Then a QML script is started from the UI, isStandalone() returns false,
but when it is started using qml or qmlscene, it returns true.
As QML scripts can now also be started with kid3-cli, they should get a
false result, when the "execute @qml script" command is called
interactively from a kid3-cli session, but a true result, if kid3-cli is
only invoking the script without an interactive session, e.g. using
kid3-cli -c "execute @qml script". The last case was not handled
correctly and is now fixed.