Fix creation of wrong editor in frame table after selection change
This issue occurred only in special situations like this:
A file has one more frame than another, e.g. a comment frame. The file
with more frames is selected, and the frame below genre (e.g. disc number)
has focus. Now the file with fewer frames is selected. Its disc number
row in the frame table has the same row number as the genre in the
previously selected file. A genre combo box is now generated in its
disc number row and populated with the previous genre value. When it
loses focus, its value is overwritten with this wrong genre value.
This happens because the call of beginRemoveRows() in
FrameTableModel::transferFrames() will trigger
FrameItemDelegate::createEditor(). Now we mark the model as temporarily
invalid and will not create an editor for the invalid frame type.