Export CSV, JSON: Only export selected folders
For the normal case, where a single file is selected and "Export CSV"
is executed from the file list context menu, nothing changes.
If any folders are selected, only the files from those folders
will be exported.
It is still possible to only export a subset of the frames by
selecting them, if this is done together with folder selection, first
the folders have to be selected and then a file. Then the frames which
shall be exported have to be selected and finally "Export CSV" invoked.
The annoying side effect of the frame selection feature that an
empty export is created just because multiple files are selected and
therefore the frame checkboxes are unchecked is now fixed, in such
a case all frames are exported. This is fixed for files having two
tags, so that it is still possible to export all frames of one tag
and skipping the other.