SelectChapterTags.qml: Support multi command kid3-cli invocation
When running SelectChapterTags.qml as an intermediate step of a
kid3-cli invocation, e.g.

kid3-cli \
  -c "execute @qml /usr/share/kid3/qml/script/SelectChapterTags.qml --deselect" \
  -c remove \
  track.mp3

the problem is that the QML script is run in stand-alone mode,
i.e. it reads the current folder and forgets about the current
selection, in this case the file track.mp3. In this example the
subsequent "remove" step operated on an empty selection thus
doing nothing.

For other use cases, stand-alone is the desired mode, e.g. using

kid3-cli -c "execute @qml '/path/to/file.qml'"

as a means to run a Kid3-QML-script without needing another
executable such as qml or qmlscene. So the current behavior
is not changed, but it is now possible to pass "--standalone=false"
to the QML script to explicitly set the mode. In the case of
SelectChapterTags.qml, running in stand-alone mode does not
make sense, to this is now switched off by overriding the
isStandalone() method in the script.

Moreover, the output
"When matching arguments for Kid3Application::tagsToFrameModels():
Too many arguments, ignoring 2", which is annoying but probably
not causing a problem is suppressed by wrapping the receiving
slot function into a function without arguments.

BUG 513005